Refactor: Decompose SidePanel into AiPanel, NotesPanel, ChatPanel
- Extracted AI logic to AiPanel.svelte
- Extracted Notes logic to NotesPanel.svelte
- Extracted Chat logic to ChatPanel.svelte
- Simplified SidePanel.svelte to handle layout and mode switching only
- Verified via svelte-check and temporary test route
